requirements.txt
Django==4.2.7
pymysql==1.1.0
Pillow==10.0.0
python-dotenv==1.0.0
reportlab==4.0.4
django-crispy-forms==2.0
crispy-bootstrap4==2023.1
django-environ==0.11.2
django-filter==23.3
django-widget-tweaks==1.5.0
django-cleanup==7.0.0
django-debug-toolbar==4.2.0
django-storages==1.14.2
boto3==1.28.62
gunicorn==21.2.0
whitenoise==6.5.0
mysqlclient==2.2.0
python-dateutil==2.8.2
crispy-bootstrap5==2023.1 